refactor(web): tighten mock api budget debt
